+++ src/components/link.tsx
@@ -1,5 +1,4 @@
 import { ArrowUpRightIcon } from 'lucide-react'
-import NextLink from 'next/link'
 import type { z } from 'zod/v4'
 import type { LinkProps, LinkVariantEnum } from '../types'
 import { cn } from '../utils'
@@ -11,13 +10,14 @@ const linkVariantClasses: Record<z.infer<typeof LinkVariantEnum>, string> = {
 
 export const Link: React.FC<LinkProps> = ({ href, children, target, variant = 'primary' }) => {
 	return (
-		<NextLink
+		<a
 			href={href}
 			target={target}
+			rel={target === '_blank' ? 'noopener noreferrer' : undefined}
 			className={cn('inline-flex items-center', variant && linkVariantClasses[variant])}
 		>
 			{children}
 			{target === '_blank' && variant !== 'ghost' && <ArrowUpRightIcon className="size-4" />}
-		</NextLink>
+		</a>
 	)
 }